proposed_build/pyproject.toml
[project]
name = "poc-self-improved-project"
version = "0.1.0"
description = "Vendor-neutral LLM ETL/ELT pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "polars>=1.0.0",
    "duckdb>=1.0.0",
    "litellm>=1.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.3.0"
]
